# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CWCM API
  version: 0.1.1-draft
  description: >
    Draft API contract baseline for Corporate Wallpaper Campaign Manager.
    This document reflects the intended backend surface required by the PRD.
    Every bearer-protected operation may return the standard 401 error response;
    the reverse proxy may return a structured 503 API_UNAVAILABLE response when
    the application server cannot be reached.
servers:
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with an assigned local or Active Directory account
      description: Case-insensitive exact assigned username (AD sAMAccountName or UPN as assigned). Active non-deleted portal assignment required before checking credentials. AD uses TLS (certificate verification defaults on; explicit LDAP_TLS_REJECT_UNAUTHORIZED=false deployment override) and pinned objectGUID; local accounts use bcrypt. No account provisioning, group-derived roles or local fallback for AD failures. Account revision validation and session issuance are atomic against portal access updates/revocation; a changed revision or serialization conflict returns generic 401 without retrying previously verified credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "401":
          description: Generic invalid credentials, unassigned, disabled, deleted or changed identity; no session issued
        "503":
          description: Authentication backend unavailable (includes LDAP/TLS/service-bind failures); no local fallback
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/session:
    get:
      tags: [Auth]
      summary: Validate and inspect the current session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out
  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Get dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Campaign"
    post:
      tags: [Campaigns]
      summary: Create campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "400":
          description: Invalid campaign, schedule conflict, or missing/deleted wallpaper
        "403":
          description: Operator or Administrator role required
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}:
    get:
      tags: [Campaigns]
      summary: Get campaign detail
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      security:
        - bearerAuth: []
      tags: [Campaigns]
      summary: Update campaign
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignRequest"
      responses:
        "400":
          description: Invalid campaign, schedule conflict, or missing/deleted wallpaper
        "403":
          description: Operator or Administrator role required
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      security:
        - bearerAuth: []
      tags: [Campaigns]
      summary: Delete campaign
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "403":
          description: Operator or Administrator role required
        "204":
          description: Campaign deleted
  /campaigns/{campaignId}/activate:
    post:
      tags: [Campaigns]
      summary: Activate campaign immediately
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "403":
          description: Operator or Administrator role required
        "200":
          description: Campaign activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/cancel:
    post:
      tags: [Campaigns]
      summary: Cancel campaign
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "403":
          description: Operator or Administrator role required
        "200":
          description: Campaign cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{campaignId}/duplicate:
    post:
      tags: [Campaigns]
      summary: Duplicate campaign as draft
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "403":
          description: Operator or Administrator role required
        "201":
          description: Campaign duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /queue:
    get:
      tags: [Queue]
      summary: Get ordered queue
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Queue state
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/QueueEntry"
                  state:
                    type: string
                    enum: [RUNNING, PAUSED]
  /queue/reorder:
    post:
      tags: [Queue]
      summary: Reorder queue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderQueueRequest"
      responses:
        "200":
          description: Queue reordered
  /queue/pause:
    post:
      tags: [Queue]
      summary: Pause queue processing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pause accepted
  /queue/resume:
    post:
      tags: [Queue]
      summary: Resume queue processing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resume accepted
  /queue/{campaignId}:
    delete:
      tags: [Queue]
      summary: Remove campaign from queue
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Queue updated after removal
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/QueueEntry"
  /deployments:
    get:
      tags: [Deployments]
      summary: List deployment history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 1-based page index for deployment history pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 250
            default: 100
          description: Maximum number of deployment history rows returned per page
      responses:
        "200":
          description: Deployment log list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeploymentLog"
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  pageCount:
                    type: integer
...
  /deployments/{deploymentId}/verify:
    post:
      tags: [Deployments]
      summary: Re-run deployment verification
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeploymentId"
      responses:
        "200":
          description: Verification accepted
  /deployments/force:
    post:
      tags: [Deployments]
      summary: Trigger manual deployment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentLog"
  /activity:
    get:
      tags: [Activity]
      summary: List activity logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Activity log list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ActivityLog"
  /settings:
    get:
      tags: [Settings]
      summary: Get runtime settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    put:
      tags: [Settings]
      summary: Update runtime settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSystemSettingsRequest"
      responses:
        "403":
          description: Administrator role required
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
  /users:
    get:
      tags: [Auth]
      summary: List users
      security:
        - bearerAuth: []
      responses:
        "403":
          description: Administrator role required
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
    post:
      tags: [Auth]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "503":
          description: AD cannot be reached or trusted while validating a new/changed AD identity
        "403":
          description: Administrator role required
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    patch:
      tags: [Auth]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "503":
          description: AD cannot be reached or trusted while validating a new/changed AD identity
        "403":
          description: Administrator role required
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [Auth]
      summary: Remove user access while preserving historical ownership
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "403":
          description: Administrator role required
        "204":
          description: User hidden from the portal, sessions revoked, and audit references preserved
        "400":
          description: Self-deletion or deletion of the last active administrator was rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scheduler/run:
    post:
      tags: [Deployments]
      summary: Trigger scheduler cycle manually
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Scheduler cycle accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  queueState:
                    type: string
                    enum: [RUNNING, PAUSED]
                  reason:
                    type: string
                    nullable: true
                  deployment:
                    allOf:
                      - $ref: "#/components/schemas/DeploymentLog"
                    nullable: true
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: >
        Protected endpoints return the standard Unauthorized response when the
        token is missing, expired, revoked, unknown, or belongs to an inactive user.
  responses:
    Unauthorized:
      description: Authentication is missing or the session is no longer valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
...
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthSession:
      type: object
      required: [token, expiresAt, user]
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    SessionStatus:
      type: object
      required: [expiresAt, user]
      properties:
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        wallpaperId:
          type: string
          format: uuid
        wallpaperTitle:
          type: string
        wallpaperImageUrl:
          type: string
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        timeZone:
          type: string
          nullable: true
          description: IANA timezone identifier used to interpret the scheduled local start and end time
        priority:
          type: integer
        status:
          type: string
          enum: [DRAFT, SCHEDULED, ACTIVE, COMPLETED, CANCELLED]
    CreateCampaignRequest:
      type: object
      required: [name, wallpaperId, priority]
      properties:
        name:
          type: string
        wallpaperId:
          type: string
          format: uuid
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        timeZone:
          type: string
          nullable: true
        priority:
          type: integer
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        wallpaperId:
          type: string
          format: uuid
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        timeZone:
          type: string
          nullable: true
        priority:
          type: integer
        status:
          type: string
          enum: [DRAFT, SCHEDULED, ACTIVE, COMPLETED, CANCELLED]
...
    DashboardSummary:
      type: object
      properties:
        currentCampaign:
          $ref: "#/components/schemas/Campaign"
        nextCampaign:
          $ref: "#/components/schemas/Campaign"
        schedulerStatus:
          type: object
          properties:
            healthy:
              type: boolean
            queueState:
              type: string
              enum: [RUNNING, PAUSED]
            intervalMinutes:
              type: integer
            lastRunAt:
              type: string
              format: date-time
              nullable: true
            nextRunAt:
              type: string
              format: date-time
              nullable: true
        deploymentStats:
          type: object
          properties:
            success:
              type: integer
            failed:
              type: integer
            warning:
              type: integer
            total:
              type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
        systemInfo:
          type: object
          properties:
            sysvolPath:
              type: string
            wallpaperFilename:
              type: string
            serverTime:
              type: string
              format: date-time
            defaultWallpaperId:
              type: string
              format: uuid
              nullable: true
            defaultWallpaperTitle:
              type: string
              nullable: true
...